Move LPad and RPad to contrib
