+++ tortoise/contrib/postgres/functions.py
@@ -1,4 +1,8 @@
+from pypika_tortoise import functions
 from pypika_tortoise.terms import Function, Term
+
+from tortoise.expressions import CombinedExpression, F
+from tortoise.functions import Function as TortoiseFunction
 
 
 class ToTsVector(Function):
@@ -37,3 +41,39 @@ class Random(Function):
 
     def __init__(self, alias=None) -> None:
         super().__init__("RANDOM", alias=alias)
+
+
+class LPad(TortoiseFunction):
+    """
+    Pads the left side of a string with a specified character to reach a certain length.
+
+    :samp:`LPad("{FIELD_NAME}", length, fill_text)`
+    """
+
+    def __init__(
+        self,
+        field: str | F | CombinedExpression | TortoiseFunction | Term,
+        length: int,
+        fill_text: str = " ",
+    ) -> None:
+        super().__init__(field, length, fill_text)
+
+    database_func = functions.LPad
+
+
+class RPad(TortoiseFunction):
+    """
+    Pads the right side of a string with a specified character to reach a certain length.
+
+    :samp:`RPad("{FIELD_NAME}", length, fill_text)`
+    """
+
+    def __init__(
+        self,
+        field: str | F | CombinedExpression | TortoiseFunction | Term,
+        length: int,
+        fill_text: str = " ",
+    ) -> None:
+        super().__init__(field, length, fill_text)
+
+    database_func = functions.RPad
